Apply theme change on settings save
Instead of only doing so on LOOT startup.

Changes to QPalette don't seem to reliably trigger UI updates, so as
part of implementing this functionality LOOT no longer uses QPalette in
its theming (mostly - in a few cases it's used for sensible defaults,
and it's unavoidable for setting the link text colour).

This means that themes are now defined using only a QSS file
(*.palette.toml is no longer used).